+++ sdk/core/azure-core/src/main/java/com/azure/core/models/CloudEventDataFormat.java
@@ -1,0 +1,23 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.core.models;
+
+/**
+ * Representation of the data format for a {@link CloudEvent}.
+ * <p>
+ * When constructing a {@link CloudEvent} this is passed to determine the serialized format of the event's data.
+ * If {@link #BYTES} is used the data will be stored as a Base64 encoded string,
+ * otherwise it will be stored as a JSON serialized object.
+ */
+public enum CloudEventDataFormat {
+    /**
+     * Bytes format.
+     */
+    BYTES,
+
+    /**
+     * JSON format.
+     */
+    JSON,
+}