+++ sdk/cosmos/azure-cosmos-spark_3-0_2-12/src/main/scala/com/azure/cosmos/spark/ItemsPartitionReader.scala
@@ -3,9 +3,8 @@
 
 package com.azure.cosmos.spark
 
-import com.azure.cosmos.CosmosClientBuilder
-import com.azure.cosmos.implementation.{CosmosClientMetadataCachesSnapshot, SparkBridgeImplementationInternal}
-import com.azure.cosmos.models.{CosmosParameterizedQuery, CosmosQueryRequestOptions}
+import com.azure.cosmos.implementation.CosmosClientMetadataCachesSnapshot
+import com.azure.cosmos.models.{CosmosParameterizedQuery, CosmosQueryRequestOptions, FeedRange}
 import com.fasterxml.jackson.databind.node.ObjectNode
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.sql.catalyst.InternalRow
@@ -18,6 +17,7 @@ import org.apache.spark.sql.types.StructType
 private case class ItemsPartitionReader
 (
   config: Map[String, String],
+  feedRange: FeedRange,
   readSchema: StructType,
   cosmosQuery: CosmosParameterizedQuery,
   cosmosClientStateHandle: Broadcast[CosmosClientMetadataCachesSnapshot]
@@ -26,6 +26,8 @@ private case class ItemsPartitionReader
 // requires making a serializable wrapper on top of SqlQuerySpec
   extends PartitionReader[InternalRow] with CosmosLoggingTrait {
   logInfo(s"Instantiated ${this.getClass.getSimpleName}")
+
+  logDebug(s"Reading from ${feedRange}")
 
   private val containerTargetConfig = CosmosContainerConfig.parseCosmosContainerConfig(config)
   private val readConfig = CosmosReadConfig.parseCosmosReadConfig(config)
@@ -38,7 +40,8 @@ private case class ItemsPartitionReader
   private lazy val iterator = cosmosAsyncContainer.queryItems(
     cosmosQuery.toSqlQuerySpec,
     new CosmosQueryRequestOptions(),
-    classOf[ObjectNode]).toIterable.iterator()
+    classOf[ObjectNode],
+    feedRange).toIterable.iterator()
 
   override def next(): Boolean = iterator.hasNext