+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/query/DocumentQueryExecutionContextBase.java
@@ -89,10 +89,13 @@ implements IDocumentQueryExecutionContext<T> {
                                                                     SqlQuerySpec querySpec,
                                                                     PartitionKeyInternal partitionKeyInternal,
                                                                     PartitionKeyRange targetRange,
-                                                                    String collectionRid) {
+                                                                    String collectionRid,
+                                                                    String throughputControlGroup) {
         RxDocumentServiceRequest request = querySpec != null
                 ? this.createQueryDocumentServiceRequest(requestHeaders, querySpec)
                 : this.createReadFeedDocumentServiceRequest(requestHeaders);
+        request.requestContext.resolvedCollectionRid = collectionRid;
+        request.throughputControlGroupName = throughputControlGroup;
 
         if (partitionKeyInternal != null) {
             request.setPartitionKeyInternal(partitionKeyInternal);
@@ -107,10 +110,13 @@ implements IDocumentQueryExecutionContext<T> {
                                                                     SqlQuerySpec querySpec,
                                                                     PartitionKeyInternal partitionKeyInternal,
                                                                     FeedRange feedRange,
-                                                                    String collectionRid) {
+                                                                    String collectionRid,
+                                                                    String throughputControlGroupName) {
         RxDocumentServiceRequest request = querySpec != null
                                                ? this.createQueryDocumentServiceRequest(requestHeaders, querySpec)
                                                : this.createReadFeedDocumentServiceRequest(requestHeaders);
+        request.requestContext.resolvedCollectionRid = collectionRid;
+        request.throughputControlGroupName = throughputControlGroupName;
 
         if (partitionKeyInternal != null) {
             feedRange = new FeedRangePartitionKeyImpl(partitionKeyInternal);