+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosChangeFeedRequestOptions.java
@@ -33,6 +33,7 @@ public final class CosmosChangeFeedRequestOptions {
     private ChangeFeedStartFromInternal startFromInternal;
     private boolean isSplitHandlingDisabled;
     private boolean quotaInfoEnabled;
+    private String throughputControlGroupName;
 
     private CosmosChangeFeedRequestOptions(
         FeedRangeInternal feedRange,
@@ -360,6 +361,7 @@ public final class CosmosChangeFeedRequestOptions {
                 CosmosChangeFeedRequestOptions.createForProcessingFromContinuation(
                     pagedFluxOptions.getRequestContinuation());
             effectiveRequestOptions.setMaxPrefetchPageCount(this.getMaxPrefetchPageCount());
+            effectiveRequestOptions.setThroughputControlGroupName(this.getThroughputControlGroupName());
         }
 
         if (pagedFluxOptions.getMaxItemCount() != null) {
@@ -406,6 +408,28 @@ public final class CosmosChangeFeedRequestOptions {
         }
 
         this.mode = ChangeFeedMode.FULL_FIDELITY;
+        return this;
+    }
+
+    /**
+     * Get the throughput control group name.
+     *
+     * @return The throughput control group name.
+     */
+    @Beta(value = Beta.SinceVersion.V4_13_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public String getThroughputControlGroupName() {
+        return this.throughputControlGroupName;
+    }
+
+    /**
+     * Set the throughput control group name.
+     *
+     * @param throughputControlGroupName The throughput control group name.
+     * @return A {@link CosmosChangeFeedRequestOptions}.
+     */
+    @Beta(value = Beta.SinceVersion.V4_13_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public CosmosChangeFeedRequestOptions setThroughputControlGroupName(String throughputControlGroupName) {
+        this.throughputControlGroupName = throughputControlGroupName;
         return this;
     }
 }