+++ sdk/spring/azure-spring-integration-servicebus/src/main/java/com/azure/spring/integration/servicebus/ServiceBusTemplate.java
@@ -42,10 +42,11 @@ public class ServiceBusTemplate<T extends ServiceBusSenderFactory> implements Se
     protected CheckpointConfig checkpointConfig = CheckpointConfig.builder()
         .checkpointMode(CheckpointMode.RECORD).build();
 
-    protected ServiceBusMessageConverter messageConverter = new ServiceBusMessageConverter();
+    protected ServiceBusMessageConverter messageConverter;
 
-    public ServiceBusTemplate(@NonNull T senderFactory) {
+    public ServiceBusTemplate(@NonNull T senderFactory, @NonNull ServiceBusMessageConverter messageConverter) {
         this.senderFactory = senderFactory;
+        this.messageConverter = messageConverter;
         LOG.info("Started ServiceBusTemplate with properties: {}", checkpointConfig);
     }