+++ sdk/spring/azure-spring-integration-servicebus/src/main/java/com/azure/spring/integration/servicebus/topic/support/ServiceBusTopicTestOperation.java
@@ -5,6 +5,7 @@ package com.azure.spring.integration.servicebus.topic.support;
 
 import com.azure.spring.integration.servicebus.ServiceBusMessageHandler;
 import com.azure.spring.integration.servicebus.ServiceBusRuntimeException;
+import com.azure.spring.integration.servicebus.converter.ServiceBusMessageConverter;
 import com.azure.spring.integration.servicebus.factory.ServiceBusTopicClientFactory;
 import com.azure.spring.integration.servicebus.topic.ServiceBusTopicTemplate;
 import com.google.common.collect.ArrayListMultimap;
@@ -29,7 +30,7 @@ public class ServiceBusTopicTestOperation extends ServiceBusTopicTemplate {
         new ConcurrentHashMap<>();
 
     public ServiceBusTopicTestOperation(ServiceBusTopicClientFactory clientFactory) {
-        super(clientFactory);
+        super(clientFactory, new ServiceBusMessageConverter());
     }
 
     @Override
@@ -47,7 +48,7 @@ public class ServiceBusTopicTestOperation extends ServiceBusTopicTemplate {
     }
 
     @Override
-    @SuppressWarnings({"rawtypes", "unchecked"})
+    @SuppressWarnings({ "rawtypes", "unchecked" })
     protected void internalSubscribe(String name, String consumerGroup, Consumer<Message<?>> consumer,
                                      Class<?> payloadType) {
         ISubscriptionClient subscriptionClient = this.senderFactory.getOrCreateSubscriptionClient(name, consumerGroup);