+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/implementation/Utility.java
@@ -8,6 +8,7 @@ import com.azure.ai.textanalytics.implementation.models.DocumentKeyPhrases;
 import com.azure.ai.textanalytics.implementation.models.DocumentStatistics;
 import com.azure.ai.textanalytics.implementation.models.EntitiesResult;
 import com.azure.ai.textanalytics.implementation.models.ErrorCodeValue;
+import com.azure.ai.textanalytics.implementation.models.ErrorResponse;
 import com.azure.ai.textanalytics.implementation.models.ErrorResponseException;
 import com.azure.ai.textanalytics.implementation.models.HealthcareResult;
 import com.azure.ai.textanalytics.implementation.models.InnerError;
@@ -45,20 +46,14 @@ import com.azure.ai.textanalytics.util.ExtractKeyPhrasesResultCollection;
 import com.azure.ai.textanalytics.util.RecognizeEntitiesResultCollection;
 import com.azure.ai.textanalytics.util.RecognizePiiEntitiesResultCollection;
 import com.azure.core.exception.HttpResponseException;
-import com.azure.core.http.HttpHeaders;
-import com.azure.core.http.HttpResponse;
-import com.azure.core.http.rest.SimpleResponse;
 import com.azure.core.util.Context;
 import com.azure.core.util.CoreUtils;
 import com.azure.core.util.IterableStream;
 import com.azure.core.util.logging.ClientLogger;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 
-import java.nio.ByteBuffer;
-import java.nio.charset.Charset;
 import java.time.Duration;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -97,62 +92,22 @@ public final class Utility {
     }
 
     /**
-     * Get a mock {@link HttpResponse} that only return status code 400.
-     *
-     * @param response A {@link SimpleResponse} with any type
-     * @return A mock {@link HttpResponse} that only return status code 400.
-     */
-    public static HttpResponse getEmptyErrorIdHttpResponse(SimpleResponse<?> response) {
-        return new HttpResponse(response.getRequest()) {
-            @Override
-            public int getStatusCode() {
-                return 400;
-            }
-
-            @Override
-            public String getHeaderValue(String s) {
-                return null;
-            }
-
-            @Override
-            public HttpHeaders getHeaders() {
-                return null;
-            }
-
-            @Override
-            public Flux<ByteBuffer> getBody() {
-                return null;
-            }
-
-            @Override
-            public Mono<byte[]> getBodyAsByteArray() {
-                return null;
-            }
-
-            @Override
-            public Mono<String> getBodyAsString() {
-                return null;
-            }
-
-            @Override
-            public Mono<String> getBodyAsString(Charset charset) {
-                return null;
-            }
-        };
-    }
-
-    /**
      * Mapping a {@link ErrorResponseException} to {@link HttpResponseException} if exist. Otherwise, return
      * original {@link Throwable}.
      *
      * @param throwable A {@link Throwable}.
      * @return A {@link HttpResponseException} or the original throwable type.
      */
-    public static Throwable mapToHttpResponseExceptionIfExist(Throwable throwable) {
+    public static Throwable mapToHttpResponseExceptionIfExists(Throwable throwable) {
         if (throwable instanceof ErrorResponseException) {
             ErrorResponseException errorException = (ErrorResponseException) throwable;
+            final ErrorResponse errorResponse = errorException.getValue();
+            com.azure.ai.textanalytics.models.TextAnalyticsError textAnalyticsError = null;
+            if (errorResponse != null && errorResponse.getError() != null) {
+                textAnalyticsError = toTextAnalyticsError(errorResponse.getError());
+            }
             return new HttpResponseException(errorException.getMessage(), errorException.getResponse(),
-                toTextAnalyticsError(errorException.getValue().getError()));
+                textAnalyticsError);
         }
         return throwable;
     }
@@ -474,24 +429,26 @@ public final class Utility {
                     new HashMap<>();
                 if (!CoreUtils.isNullOrEmpty(documentEntities.getRelations())) {
                     documentEntities.getRelations().forEach(healthcareRelation -> {
-                        final HealthcareEntity targetEntity =
-                            healthcareEntities.get(getHealthcareEntityIndex(healthcareRelation.getTarget()));
-                        final HealthcareEntity sourceEntity =
-                            healthcareEntities.get(getHealthcareEntityIndex(healthcareRelation.getSource()));
-                        final HealthcareEntityRelationType relationType =
-                            HealthcareEntityRelationType.fromString(healthcareRelation.getRelationType());
+                        // TODO: bidirectional relation is gone, do it in another PR
+                        // issue: https://github.com/Azure/azure-sdk-for-java/issues/19180
+//                        final HealthcareEntity targetEntity =
+//                            healthcareEntities.get(getHealthcareEntityIndex(healthcareRelation.getTarget()));
+//                        final HealthcareEntity sourceEntity =
+//                            healthcareEntities.get(getHealthcareEntityIndex(healthcareRelation.getSource()));
+//                        final HealthcareEntityRelationType relationType =
+//                            HealthcareEntityRelationType.fromString(healthcareRelation.getRelationType());
 
-                        final Map<HealthcareEntity, HealthcareEntityRelationType> targetRelatedEntity =
-                            entityRelationMap.getOrDefault(targetEntity, new HashMap<>());
-                        targetRelatedEntity.putIfAbsent(sourceEntity, relationType);
-                        entityRelationMap.putIfAbsent(targetEntity, targetRelatedEntity);
-
-                        if (healthcareRelation.isBidirectional()) {
-                            final Map<HealthcareEntity, HealthcareEntityRelationType> sourceRelatedEntity =
-                                entityRelationMap.getOrDefault(sourceEntity, new HashMap<>());
-                            sourceRelatedEntity.putIfAbsent(targetEntity, relationType);
-                            entityRelationMap.putIfAbsent(sourceEntity, sourceRelatedEntity);
-                        }
+//                        final Map<HealthcareEntity, HealthcareEntityRelationType> targetRelatedEntity =
+//                            entityRelationMap.getOrDefault(targetEntity, new HashMap<>());
+//                        targetRelatedEntity.putIfAbsent(sourceEntity, relationType);
+//                        entityRelationMap.putIfAbsent(targetEntity, targetRelatedEntity);
+//
+//                        if (healthcareRelation.isBidirectional()) {
+//                            final Map<HealthcareEntity, HealthcareEntityRelationType> sourceRelatedEntity =
+//                                entityRelationMap.getOrDefault(sourceEntity, new HashMap<>());
+//                            sourceRelatedEntity.putIfAbsent(targetEntity, relationType);
+//                            entityRelationMap.putIfAbsent(sourceEntity, sourceRelatedEntity);
+//                        }
                     });
                 }
                 healthcareEntities.forEach(healthcareEntity -> {
@@ -523,24 +480,24 @@ public final class Utility {
         return IterableStream.of(analyzeHealthcareEntitiesResults);
     }
 
-    /**
-     * Helper function that parse healthcare entity index from the given entity reference string.
-     * The entity reference format is "#/results/documents/0/entities/3".
-     *
-     * @param entityReference the given healthcare entity reference string.
-     *
-     * @return the healthcare entity index.
-     */
-    private static Integer getHealthcareEntityIndex(String entityReference) {
-        if (!CoreUtils.isNullOrEmpty(entityReference)) {
-            int lastIndex = entityReference.lastIndexOf('/');
-            if (lastIndex != -1) {
-                return Integer.parseInt(entityReference.substring(lastIndex + 1));
-            }
-        }
-        throw LOGGER.logExceptionAsError(
-            new RuntimeException("Failed to parse healthcare entity index from: " + entityReference));
-    }
+//    /**
+//     * Helper function that parse healthcare entity index from the given entity reference string.
+//     * The entity reference format is "#/results/documents/0/entities/3".
+//     *
+//     * @param entityReference the given healthcare entity reference string.
+//     *
+//     * @return the healthcare entity index.
+//     */
+//    private static Integer getHealthcareEntityIndex(String entityReference) {
+//        if (!CoreUtils.isNullOrEmpty(entityReference)) {
+//            int lastIndex = entityReference.lastIndexOf('/');
+//            if (lastIndex != -1) {
+//                return Integer.parseInt(entityReference.substring(lastIndex + 1));
+//            }
+//        }
+//        throw LOGGER.logExceptionAsError(
+//            new RuntimeException("Failed to parse healthcare entity index from: " + entityReference));
+//    }
 
     /**
      * Transfer {@link com.azure.ai.textanalytics.models.StringIndexType} into auto-generated {@link StringIndexType}.
@@ -566,5 +523,21 @@ public final class Utility {
      */
     public static Context getNotNullContext(Context context) {
         return context == null ? Context.NONE : context;
+    }
+
+    /**
+     * Helper function which retrieves the size of an {@link Iterable}.
+     *
+     * @param documents The iterable of documents.
+     * @return Count of documents in the iterable.
+     */
+    public static int getDocumentCount(Iterable<?> documents) {
+        if (documents instanceof Collection) {
+            return ((Collection<?>) documents).size();
+        } else {
+            final int[] count = new int[] { 0 };
+            documents.forEach(ignored -> count[0] += 1);
+            return count[0];
+        }
     }
 }